+++ Form/Extension/FormFlowHiddenFieldExtension.php
@@ -9,10 +9,11 @@ use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 
 /**
  * @author Konstantin Myakshin <koc-dp@yandex.ru>
+ * @author Christian Raue <christian.raue@gmail.com>
  * @copyright 2011-2013 Christian Raue
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
-class FormFlowStepFieldExtension extends AbstractTypeExtension {
+class FormFlowHiddenFieldExtension extends AbstractTypeExtension {
 
 	/**
 	 * {@inheritDoc}
@@ -26,6 +27,7 @@ class FormFlowStepFieldExtension extends AbstractTypeExtension {
 	 */
 	public function setDefaultOptions(OptionsResolverInterface $resolver) {
 		$resolver->setOptional(array(
+			'flow_instance_key',
 			'flow_step_key',
 		));
 	}
@@ -34,6 +36,11 @@ class FormFlowStepFieldExtension extends AbstractTypeExtension {
 	 * {@inheritDoc}
 	 */
 	public function finishView(FormView $view, FormInterface $form, array $options) {
+		if (array_key_exists('flow_instance_key', $options) && $view->vars['name'] === $options['flow_instance_key']) {
+			$view->vars['value'] = $options['data'];
+			$view->vars['full_name'] = $options['flow_instance_key'];
+		}
+
 		if (array_key_exists('flow_step_key', $options) && $view->vars['name'] === $options['flow_step_key']) {
 			$view->vars['value'] = $options['data'];
 			$view->vars['full_name'] = $options['flow_step_key'];